Changed higherKindedBounds to not split the original bound type.
But now another problem arises:

Say our bound is of the form

     Tycon { type _$hk0 >: L <: H }

But now any map over this bound will replace the _$hk0 name by the corresponding type parameter in Tycon.
This happens in derivedRefiendType. This is of course not what we want, but it seems there's too little information available to prevent it.

In light of this the original scheme for treating higherkinded types based on anding with hkTraits does not look so bad. Sor, for the time being this branch is aborted.