support for apply in package objects
Previously, it was not possible to write `packageName(args)` even if the
associated package object had an `apply` method, however that's what we would like
to have for https://github.com/odersky/sips/blob/gh-pages/drafts/inlinemeta.md.

It turns out that enabling this functionality is really easy - the `checkValue`
guard that used to fail on any `sym is Package` can be changed to account
for `apply` methods if we're typechecking against a FunProto prototype.