new type: TypeRef with a fixed sym and fixed info.
Linker uses it to represent types that normally would never have
escaped local scope(and thus have NoPrefix as prefix) that do
escape in Linker.

Eg, GenTraversableFactory has inner class GenericCanBuildFrom,
and anonymous instance of this class in field ReusableCBFInstance.

Linker needs to somehow see that the anonymous instance instantiated
in List$.ReusableCBFInstance is a subtype of List$.GenTraversableFactory
Thus it needs a valid prefix to participate in asSeenFrom and substThis.